requirements.txt
flask
pillow
rembg